+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/apiModel/response/MessageRoomResponse.kt
@@ -1,4 +1,4 @@
-package com.emmsale.data.messageRoom.dto
+package com.emmsale.data.apiModel.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
