+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/message/mapper/MessageMapper.kt
@@ -1,7 +1,7 @@
 package com.emmsale.data.message.mapper
 
-import com.emmsale.data.message.Message
-import com.emmsale.data.messageRoom.dto.MessageResponse
+import com.emmsale.data.apiModel.response.MessageResponse
+import com.emmsale.data.model.Message
 
 fun List<MessageResponse>.toData(): List<Message> = map { it.toData() }
