+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/model/Message.kt
@@ -1,4 +1,4 @@
-package com.emmsale.data.message
+package com.emmsale.data.model
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
