+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/repository/interfaces/MessageRoomRepository.kt
@@ -1,7 +1,8 @@
-package com.emmsale.data.messageRoom
+package com.emmsale.data.repository.interfaces
 
 import com.emmsale.data.common.callAdapter.ApiResponse
-import com.emmsale.data.message.Message
+import com.emmsale.data.model.Message
+import com.emmsale.data.model.MessageRoom
 
 interface MessageRoomRepository {
     suspend fun getMessageRooms(
