+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/messageList/uistate/OtherMessageUiState.kt
@@ -1,7 +1,7 @@
 package com.emmsale.presentation.ui.messageList.uistate
 
-import com.emmsale.data.message.Message
 import com.emmsale.data.model.Member
+import com.emmsale.data.model.Message
 import java.time.LocalDateTime
 
 class OtherMessageUiState(
