+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/messageRoomList/uistate/MemberRoomListUiState.kt
@@ -1,6 +1,6 @@
 package com.emmsale.presentation.ui.messageRoomList.uistate
 
-import com.emmsale.data.messageRoom.MessageRoom
+import com.emmsale.data.model.MessageRoom
 import com.emmsale.presentation.common.FetchResult
 import com.emmsale.presentation.common.FetchResultUiState
