+++ android/2023-emmsale/app/src/main/res/layout/item_message_room.xml
@@ -13,7 +13,7 @@
 
         <variable
             name="messageRoom"
-            type="com.emmsale.data.messageRoom.MessageRoom" />
+            type="com.emmsale.data.model.MessageRoom" />
 
         <variable
             name="onMessageRoomClick"
